Flytter henting av organisasjoner og tilgangsstyring i egen provider, og tar i bruk context
