+++ package.json
@@ -99,7 +99,7 @@
     "react-transition-group": "1.x",
     "regenerator-runtime": "^0.13.5",
     "serialize-javascript": "^2.1.1",
-    "typescript": "3.7.2",
+    "typescript": "3.8.3",
     "underscore.string": "^3.3.5",
     "unorm": "^1.6.0",
     "whatwg-fetch": "^3.3.1",
