+++ src/App/MineAnsatte/HovedBanner/HovedBanner.tsx
@@ -1,18 +1,20 @@
-import React, { FunctionComponent } from 'react';
+import React, { FunctionComponent, useContext } from 'react';
 import { withRouter, RouteComponentProps } from 'react-router';
 import Bedriftsmeny from '@navikt/bedriftsmeny';
 import '@navikt/bedriftsmeny/lib/bedriftsmeny.css';
 import { Organisasjon, tomaAltinnOrganisasjon } from '../../Objekter/OrganisasjonFraAltinn';
 import { nullStillSorteringIUrlParametere } from '../urlFunksjoner';
+import { OrganisasjonerOgTilgangerContext } from '../../OrganisasjonerOgTilgangerProvider';
 
 interface Props extends RouteComponentProps {
-    byttOrganisasjon: (org: Organisasjon) => void;
-    organisasjoner: Organisasjon[] | undefined;
+    hentOgSetAntallOgArbeidsforhold: (organisasjon: Organisasjon, erTidligereArbeidsforhold: boolean) => void;
+    abortTidligereRequests: () => void;
     setEndringIUrlAlert: (endret: string) => void;
-    valgtAktivOrganisasjon: Organisasjon;
 }
 
 const Banner: FunctionComponent<Props> = props => {
+    const { organisasjonerFraAltinnMedTilgang, valgtAktivOrganisasjon, setValgtAktivOrganisasjon, organisasjonerFraAltinn, setTilgangTilTidligereArbeidsforhold } = useContext(OrganisasjonerOgTilgangerContext);
+
     const { history } = props;
     const naVærendeUrl = new URL(window.location.href);
     const erPåEnkeltArbeidsforhold = naVærendeUrl.href.includes('/enkeltarbeidsforhold');
@@ -25,12 +27,27 @@ const Banner: FunctionComponent<Props> = props => {
     };
 
     const sjekkAtManBytterBedriftIkkeVedRefresh = () => {
-        return props.valgtAktivOrganisasjon !== tomaAltinnOrganisasjon;
+        return valgtAktivOrganisasjon !== tomaAltinnOrganisasjon;
+    };
+
+    const harTilgang = (orgnr: string): number | undefined => {
+        return organisasjonerFraAltinnMedTilgang?.filter(org => org.OrganizationNumber === orgnr).length;
     };
 
     const onOrganisasjonChange = (organisasjon?: Organisasjon) => {
         if (organisasjon) {
-            props.byttOrganisasjon(organisasjon);
+            setValgtAktivOrganisasjon(organisasjon);
+            if (harTilgang(organisasjon.ParentOrganizationNumber)) {
+                setTilgangTilTidligereArbeidsforhold(true);
+            } else {
+                setTilgangTilTidligereArbeidsforhold(false);
+            }
+            props.abortTidligereRequests();
+
+            if (organisasjon.OrganizationNumber && harTilgang(organisasjon.OrganizationNumber)) {
+                props.hentOgSetAntallOgArbeidsforhold(organisasjon, false);
+            }
+
             if (sjekkAtManBytterBedriftIkkeVedRefresh()) {
                 history.replace(nullStillSorteringIUrlParametere());
                 erPåEnkeltArbeidsforhold && redirectTilListeVisning();
@@ -41,7 +58,7 @@ const Banner: FunctionComponent<Props> = props => {
     };
 
     const sidetittel = erPåTidligereArbeidsforhold ? 'Tidligere arbeidsforhold' : 'Arbeidsforhold';
-    const organisasjonerIBedriftsmenyen = erPåTidligereArbeidsforhold ? [] : props.organisasjoner;
+    const organisasjonerIBedriftsmenyen = erPåTidligereArbeidsforhold ? [] : organisasjonerFraAltinn;
 
     return (
         <div className="hovebanner">
